Implement packages list
