+++ src/main/resources/application.properties
@@ -3,4 +3,4 @@ spring.datasource.url=jdbc:postgresql://localhost:5432/glvd
 spring.datasource.username=glvd
 spring.datasource.password=glvd
 spring.sql.init.mode=never
-jakarta.persistence.query.timeout=5000
+jakarta.persistence.query.timeout=7000
