+++ src/test/java/io/gardenlinux/glvd/GlvdControllerTest.java
@@ -42,7 +42,8 @@ class GlvdControllerTest {
 
     @Container
     @ServiceConnection
-    static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>(glvdPostgresImage).withDatabaseName("glvd")
+    static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>(glvdPostgresImage)
+            .withDatabaseName("glvd")
             .withUsername("glvd").withPassword("glvd");
 
     @Autowired
@@ -100,7 +101,7 @@ class GlvdControllerTest {
                 .filter(document("getCveForDistro",
                         preprocessRequest(modifyUris().scheme("https").host("glvd.gardenlinux.io").removePort()),
                         preprocessResponse(prettyPrint())))
-                .when().port(this.port).get("/v1/cves/debian_linux/bookworm")
+                .when().port(this.port).get("/v1/cves/gardenlinux/1592")
 				.then().statusCode(HttpStatus.SC_OK);
     }
 
@@ -110,7 +111,7 @@ class GlvdControllerTest {
                 .filter(document("getCveForDistroByVersion",
                         preprocessRequest(modifyUris().scheme("https").host("glvd.gardenlinux.io").removePort()),
                         preprocessResponse(prettyPrint())))
-                .when().port(this.port).get("/v1/cves/debian_linux/version/12")
+                .when().port(this.port).get("/v1/cves/gardenlinux/version/1592.0")
                 .then().statusCode(HttpStatus.SC_OK);
     }
 
@@ -120,7 +121,7 @@ class GlvdControllerTest {
                 .filter(document("getCveForPackages",
                         preprocessRequest(modifyUris().scheme("https").host("glvd.gardenlinux.io").removePort()),
                         preprocessResponse(prettyPrint())))
-                .when().port(this.port).get("/v1/cves/debian_linux/bookworm/packages/dav1d,firefox-esr")
+                .when().port(this.port).get("/v1/cves/gardenlinux/1592/packages/crun,vim")
                 .then().statusCode(HttpStatus.SC_OK);
     }
 
@@ -130,7 +131,7 @@ class GlvdControllerTest {
                 .filter(document("getCveForPackagesByDistroVersion",
                         preprocessRequest(modifyUris().scheme("https").host("glvd.gardenlinux.io").removePort()),
                         preprocessResponse(prettyPrint())))
-                .when().port(this.port).get("/v1/cves/debian_linux/version/12/packages/dav1d,firefox-esr")
+                .when().port(this.port).get("/v1/cves/gardenlinux/version/1592.0/packages/crun,vim")
                 .then().statusCode(HttpStatus.SC_OK);
     }
 
@@ -142,6 +143,16 @@ class GlvdControllerTest {
                         preprocessResponse(prettyPrint())))
                 .when().port(this.port).get("/readiness")
 				.then().statusCode(HttpStatus.SC_OK).body("dbCheck", containsString("true"));
+    }
+
+    @Test
+    public void shouldGetPackagesForDistro() {
+        given(this.spec).accept("application/json")
+                .filter(document("getPackages",
+                        preprocessRequest(modifyUris().scheme("https").host("glvd.gardenlinux.io").removePort()),
+                        preprocessResponse(prettyPrint())))
+                .when().port(this.port).get("/v1/packages/1592.0")
+                .then().statusCode(200);
     }
 
 }
